Rename truffle-contract-schema to @truffle/contract-schema
