+++ packages/contract-schema/.travis.yml
