+++ packages/contract-schema/network-object.spec.md
