+++ packages/contract-schema/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "truffle-contract-schema",
+  "name": "@truffle/contract-schema",
   "version": "3.0.13",
   "description": "JSON schema for contract artifacts",
   "main": "index.js",
@@ -8,7 +8,7 @@
     "build": "cd spec && json2ts -i contract-object.spec.json -o ./index.d.ts",
     "test": "mocha"
   },
-  "repository": "https://github.com/trufflesuite/truffle/tree/master/packages/truffle-contract-schema",
+  "repository": "https://github.com/trufflesuite/truffle/tree/develop/packages/contract-schema",
   "keywords": [
     "ethereum",
     "json",
@@ -16,12 +16,12 @@
     "contract",
     "artifacts"
   ],
-  "author": "Tim Coulter <tim.coulter@consensys.net>",
+  "author": "Tim Coulter <tim@trufflesuite.com>",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/trufflesuite/truffle/issues"
   },
-  "homepage": "https://github.com/trufflesuite/truffle/blob/develop/packages/truffle-contract-schema#readme",
+  "homepage": "https://github.com/trufflesuite/truffle/tree/develop/packages/contract-schema#readme",
   "dependencies": {
     "ajv": "^6.10.0",
     "crypto-js": "^3.1.9-1",
