+++ packages/contract-schema/spec/contract-object.spec.json
