+++ packages/contract-schema/spec/network-object.spec.json
