+++ packages/contract-schema/test/MetaCoin.json
