+++ packages/truffle-artifactor/test/contracts.js
@@ -1,7 +1,7 @@
 const assert = require("chai").assert;
 const Artifactor = require("../");
 const contract = require("truffle-contract");
-const Schema = require("truffle-contract-schema");
+const Schema = require("@truffle/contract-schema");
 const temp = require("temp").track();
 const path = require("path");
 const fs = require("fs");
