+++ packages/truffle-contract/index.js
@@ -1,4 +1,4 @@
-const Schema = require("truffle-contract-schema");
+const Schema = require("@truffle/contract-schema");
 const Contract = require("./lib/contract");
 const truffleContractVersion = require("./package.json").version;
