+++ packages/truffle-decoder/package.json
@@ -45,7 +45,7 @@
     "@types/utf8": "^2.1.6",
     "@types/web3": "^1.0.5",
     "json-schema-to-typescript": "^5.5.0",
-    "truffle-contract-schema": "^3.0.13",
+    "@truffle/contract-schema": "^3.0.13",
     "typescript": "^3.1.3"
   },
   "keywords": [
