+++ packages/truffle-external-compile/package.json
@@ -16,7 +16,7 @@
   "dependencies": {
     "debug": "^4.1.0",
     "glob": "^7.1.2",
-    "truffle-contract-schema": "^3.0.13",
+    "@truffle/contract-schema": "^3.0.13",
     "@truffle/expect": "^0.0.9",
     "web3-utils": "1.2.1"
   },
